==> WT_RPSA_Analysis/scripts/03_Grouping_WT_at_rep_pos.py <==
"""
Groups WT RPSA carrying haplotypes by their sequence at the three union set representative positions

Requirements:
- pandas==2.1.1

"""

import pandas as pd
from pathlib import Path
from path_utils import get_workflow_dir


def main():
    """
    Subset WT haplotype sequences to specific representative positions and export the result.

    The function:
    - Reads in the WT haplotypes CSV file generated by script 01.
    - Subsets the haplotypes to three specific positions obtained by fine-mapping the eQTLs of RPSA
    - Exports the subsetted DataFrame to a new CSV file in the results directory.

    """

    workflow_dir = get_workflow_dir()

    #read in WT RPSA carrying haplotype sequences
    WT_path = workflow_dir / 'data' / 'Distance_Haplotype_Intermediate_Files' / 'WT_Haplotypes.csv'
    WT_seq = pd.read_csv(WT_path, sep= ",")

    #subset to representative positions US1: 39406645 US2: 39415034 US3: 39407090
    Rep_Pos = [39406645, 39415034, 39407090]

    WT_RepPos = WT_seq[WT_seq['POS'].isin(Rep_Pos)]
    output_dir = workflow_dir / 'Results' / 'WT_RPSA_Grouping'
    WT_RepPos.to_csv(output_dir / "WT_RPSA_Groups.csv", sep=",", index=False)


if __name__ == "__main__":
    main()
